moved config to separate module
